finished "picture frame" tracker  and more city
the picture frame tracker is totally done, there are 8 items to collect, grabbing them plays a sound and makes the item disappear and then it's corresponding picture frame gets populated with a crude drawing of the item. also added lots of environment to the city scape area, namely snow hills and snowbanks. next up will be getting at least one more building variant and to add the zipline between buildings.